Get role session name from CI job name
Also remove `--var` args - they're not used anywhere in the wild.

JIRA: PLAT-945
