+++ test/test_integration_cli_lambda.py
@@ -23,6 +23,7 @@ class TestReleaseCLI(unittest.TestCase):
     @patch('cdflow_commands.release.check_call')
     @patch('cdflow_commands.release.make_archive')
     @patch('cdflow_commands.release.open', new_callable=mock_open, create=True)
+    @patch('cdflow_commands.cli.os')
     @patch('cdflow_commands.cli.rmtree')
     @patch('cdflow_commands.cli.Session')
     @patch('cdflow_commands.config.Session')
@@ -30,8 +31,9 @@ class TestReleaseCLI(unittest.TestCase):
     @patch('cdflow_commands.config.check_output')
     def test_release_package_is_created(
         self, check_output, mock_open_config, Session_from_config,
-        Session_from_cli, rmtree, mock_open_release, make_archive, check_call,
-        copytree, S3BucketFactory, mock_os, ZipFile, check_output_cli,
+        Session_from_cli, rmtree, mock_os, mock_open_release, make_archive,
+        check_call, copytree, S3BucketFactory, mock_os_lambda, ZipFile,
+        check_output_cli,
     ):
         # Given
         mock_metadata_file = MagicMock(spec=TextIOWrapper)
